+++ include/glaze/json/jmespath.hpp
@@ -15,129 +15,6 @@
 
 namespace glz
 {
-   // TODO: GCC 12 lacks constexpr std::from_chars
-   // Remove this code when dropping GCC 12 support
-   namespace detail
-   {
-      struct from_chars_result
-      {
-         const char* ptr;
-         std::errc ec;
-      };
-
-      inline constexpr int char_to_digit(char c) noexcept
-      {
-         if (c >= '0' && c <= '9') return c - '0';
-         if (c >= 'a' && c <= 'z') return c - 'a' + 10;
-         if (c >= 'A' && c <= 'Z') return c - 'A' + 10;
-         return -1;
-      }
-
-      template <class I>
-      constexpr from_chars_result from_chars(const char* first, const char* last, I& value, int base = 10)
-      {
-         from_chars_result result{first, std::errc{}};
-
-         // Basic validation of base
-         if (base < 2 || base > 36) {
-            // Not standard behavior to check base validity here, but let's return invalid_argument
-            result.ec = std::errc::invalid_argument;
-            return result;
-         }
-
-         using U = std::make_unsigned_t<I>;
-         constexpr bool is_signed = std::is_signed<I>::value;
-         constexpr U umax = (std::numeric_limits<U>::max)();
-
-         if (first == last) {
-            // Empty range
-            result.ec = std::errc::invalid_argument;
-            return result;
-         }
-
-         bool negative = false;
-         // Check for sign only if signed type
-         if constexpr (is_signed) {
-            if (*first == '-') {
-               negative = true;
-               ++first;
-            }
-            else if (*first == '+') {
-               ++first;
-            }
-         }
-
-         if (first == last) {
-            // After sign there's nothing
-            result.ec = std::errc::invalid_argument;
-            return result;
-         }
-
-         U acc = 0;
-         bool any = false;
-
-         // We'll do overflow checking as we parse
-         // For accumulation: acc * base + digit
-         // Overflow if acc > (umax - digit)/base
-
-         while (first != last) {
-            int d = char_to_digit(*first);
-            if (d < 0 || d >= base) break;
-
-            // Check overflow before multiplying/adding
-            if (acc > (umax - static_cast<U>(d)) / static_cast<U>(base)) {
-               // Overflow
-               result.ec = std::errc::result_out_of_range;
-               // We still move ptr to the last valid digit parsed
-               result.ptr = first;
-               // No need to parse further; we know it's out of range.
-               return result;
-            }
-
-            acc = acc * base + static_cast<U>(d);
-            any = true;
-            ++first;
-         }
-
-         if (!any) {
-            // No digits parsed
-            result.ec = std::errc::invalid_argument;
-            return result;
-         }
-
-         // If signed and negative, check if result fits
-         if constexpr (is_signed) {
-            using S = std::make_signed_t<U>;
-            // The largest magnitude we can represent in a negative value is (max + 1)
-            // since -(min()) = max() + 1.
-            U limit = static_cast<U>(std::numeric_limits<I>::max()) + 1U;
-            if (negative) {
-               if (acc > limit) {
-                  result.ec = std::errc::result_out_of_range;
-                  result.ptr = first;
-                  return result;
-               }
-               value = static_cast<I>(0 - static_cast<S>(acc));
-            }
-            else {
-               if (acc > static_cast<U>(std::numeric_limits<I>::max())) {
-                  result.ec = std::errc::result_out_of_range;
-                  result.ptr = first;
-                  return result;
-               }
-               value = static_cast<I>(acc);
-            }
-         }
-         else {
-            // Unsigned type
-            value = acc;
-         }
-
-         result.ptr = first;
-         return result;
-      }
-   }
-
    namespace jmespath
    {
       enum struct tokenization_error {
